% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_color_pal_cinema.R
\name{extract_cpc_pal}
\alias{extract_cpc_pal}
\title{Extract color palette from @colorpalette.cinema image}
\usage{
extract_cpc_pal(fp)
}
\arguments{
\item{fp}{The path to a @colorpalette.cinema image}
}
\description{
Extract color palette from @colorpalette.cinema image
}
