% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_from_post.R
\name{pal_from_post}
\alias{pal_from_post}
\title{Extract Palettes from an Instagram post}
\usage{
pal_from_post(post_url)
}
\arguments{
\item{post_url}{The url of the instagram post.}
}
\description{
Provide a \@colorpalette.cinema Instagram post url and extract the color paletts from the posts.
}
\examples{
pal_from_post("https://www.instagram.com/p/CMC26FaHbqP/")
}
