% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casting.R, R/cpcinema.R
\name{vec_ptype2.cpcinema}
\alias{vec_ptype2.cpcinema}
\alias{vec_cast.cpcinema}
\alias{vctrs-cpcinema}
\title{Internal vctrs methods}
\usage{
\method{vec_ptype2}{cpcinema}(x, y, ...)

\method{vec_cast}{cpcinema}(x, to, ...)
}
\description{
Internal vctrs methods
}
\keyword{internal}
