% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{color_palette}
\alias{color_palette}
\alias{is_color_palette}
\title{Construct a new color palette}
\usage{
color_palette(pal, n, type = c("discrete", "continuous"))

is_color_palette(pal)
}
\arguments{
\item{pal}{A character vector of color hex codes.}

\item{n}{(optional) number to indicate how many values to include in the color palette.}
}
\description{
Construct a new color palette
}
\note{
If `type` is set to `"discrete"`, `n` cannot be greater than the number of unique values in `pal`. `n` can be greater than number of unqiue values in `pal` when `type = "continuous"`. In the latter case, colors are interpolated with `grDevices::colorRampPalette()`.
}
