% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{new_palette}
\alias{new_palette}
\title{Color palette constructor and validator}
\usage{
new_palette(pal, n, type = c("discrete", "continuous"), ...)
}
\description{
Creates a new color_palette object. \code{new_palette()} is much more strict than \code{color_palette()} and is intended for use in development.
}
